+++ src/utils/hisWritePayload.js
@@ -4,7 +4,7 @@ class HisWritePayload {
     }
 
     /**
-     * Add data to the hisWrite payload.
+     * Add data to the hisWrite payload. If the unit is present on val attribute, it will be stripped.
      * @param id Entity id with Haystack prefix applied.
      * @param data Data of the format of [{ts: String, val: String}] with the Haystack prefix applied.
      */
@@ -21,7 +21,15 @@ class HisWritePayload {
                 this.payload[ts] = {};
             }
 
-            this.payload[ts][id] = val;
+            let parsedVal = val;
+            if (val.substring(0, 2) === "n:" && val.includes(" ")) {
+                const index= val.indexOf(" ");
+                if (index !== -1) {
+                    // val has unit applied, remove
+                    parsedVal = val.substring(0, val.indexOf(" "));
+                }
+            }
+            this.payload[ts][id] = parsedVal;
         }
     }