+++ test/client/functions/batch/hisDelete.js
@@ -26,6 +26,21 @@ const HIS_READ_LARGE_ENTITIES = {
 };
 const TIME_START = new Date(0);
 const TIME_END = new Date();
+const HIS_READ_TIME_SERIES_MIX = {
+    data: require("./files/hisRead_mix.json"),
+    ids: [
+        "06037f10-bc06-4ddd-a58e-3fe2a975e830",
+        "0e2578db-90a3-4454-95d3-38565f26c0f6",
+        "105b1cd6-cb74-449e-9389-50165328a9e3",
+        "1a669067-d59b-4e29-b564-9c84e3784aeb",
+        "2fefaf25-555a-4184-94bc-abae4ec8d012",
+        "37cfffce-fccc-454d-b5ee-cd7999ace8bf",
+        "40685948-db22-414b-9e2e-530fc16cd576",
+        "439bec63-2903-40fa-9547-99096173c92f",
+        "68502133-c34f-43a0-8875-7ddd6f6311bb",
+        "7b50c5ad-13fd-4a17-a1d2-c589de55cce3"
+    ]
+};
 
 describe("client", () => {
     let ws, http, log;
@@ -56,11 +71,49 @@ describe("client", () => {
     });
 
     describe("batch.hisDelete", () => {
+        describe("time range batch behaviour", () => {
+            describe("mix time series ranges", () => {
+                beforeEach(() => {
+                    ws.batch.hisRead = sinon.stub().callsFake(() => {
+                        return {
+                            success: HIS_READ_TIME_SERIES_MIX.data,
+                            errors: []
+                        };
+                    });
+                });
+
+                it("should create all time ranges", async () => {
+                    await ws.batch.hisDelete(HIS_READ_TIME_SERIES_MIX.ids, TIME_START, TIME_END);
+                    expect(ws.hisDelete.callCount).to.equal(8);
+                    const expectedRanges = [
+                        "s:2022-12-17T16:30:00.000Z,2022-12-23T12:25:00.001Z",
+                        "s:2022-12-23T12:25:00.001Z,2022-12-29T10:55:00.001Z",
+                        "s:2022-12-29T10:55:00.001Z,2023-01-04T09:25:00.001Z",
+                        "s:2023-01-04T09:25:00.001Z,2023-01-10T07:55:00.001Z",
+                        "s:2023-01-10T07:55:00.001Z,2023-01-16T06:25:00.001Z",
+                        "s:2023-01-16T06:25:00.001Z,2023-01-22T04:55:00.001Z",
+                        "s:2023-01-22T04:55:00.001Z,2023-01-28T03:25:00.001Z",
+                        "s:2023-01-28T03:25:00.001Z,2023-04-30T14:00:00.001Z"
+                    ];
+                    for (let i = 0; i < ws.hisDelete.callCount; i++) {
+                        expect(ws.hisDelete.args[i]).to.eql([
+                            HIS_READ_TIME_SERIES_MIX.ids, expectedRanges[i]
+                        ], `index=${i}`);
+                    }
+                });
+            });
+        });
+
         describe("no options specified", () => {
             describe("time series rows", () => {
                 describe("payload smaller than default batch size of 100", () => {
                     it("should only make 1 request", async () => {
-                        ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                        ws.batch.hisRead = sinon.stub().callsFake(() => {
+                            return {
+                                success: HIS_READ_SMALL_TIME_SERIES.data,
+                                errors: []
+                            }
+                        });
                         await ws.batch.hisDelete(
                             HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END);
                         expect(ws.hisDelete.calledOnce).to.be.true;
@@ -72,7 +125,12 @@ describe("client", () => {
 
                     describe("payload greater than default batch size of 100", () => {
                         it("should make more than 1 request", async () => {
-                            ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_LARGE_TIME_SERIES.data);
+                            ws.batch.hisRead = sinon.stub().callsFake(() => {
+                                return {
+                                    success: HIS_READ_LARGE_TIME_SERIES.data,
+                                    errors: []
+                                };
+                            });
                             await ws.batch.hisDelete(
                                 HIS_READ_LARGE_TIME_SERIES.ids, TIME_START, TIME_END);
                             expect(ws.hisDelete.calledTwice).to.be.true;
@@ -91,7 +149,12 @@ describe("client", () => {
                 describe("entities", () => {
                     describe("entities smaller than maximum of 100", () => {
                         it("should only make 1 request", async () => {
-                            ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                            ws.batch.hisRead = sinon.stub().callsFake(() => {
+                                return {
+                                    success: HIS_READ_SMALL_TIME_SERIES.data,
+                                    errors: []
+                                };
+                            });
                             await ws.batch.hisDelete(
                                 HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END);
                             expect(ws.hisDelete.calledOnce).to.be.true;
@@ -104,7 +167,12 @@ describe("client", () => {
 
                     describe("entities greater than maximum of 100", () => {
                         it("should make more than 1 request", async () => {
-                            ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_LARGE_ENTITIES.data);
+                            ws.batch.hisRead = sinon.stub().callsFake(() => {
+                                return {
+                                    success: HIS_READ_LARGE_ENTITIES.data,
+                                    errors: []
+                                };
+                            });
                             await ws.batch.hisDelete(
                                 HIS_READ_LARGE_ENTITIES.ids, TIME_START, TIME_END);
                             expect(ws.hisDelete.calledTwice).to.be.true;
@@ -125,7 +193,12 @@ describe("client", () => {
         describe("option batchSize", () => {
             describe("delete data smaller than batchSize", () => {
                 it("should make 1 request", async () => {
-                    ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                    ws.batch.hisRead = sinon.stub().callsFake(() => {
+                        return {
+                            success: HIS_READ_SMALL_TIME_SERIES.data,
+                            errors: HIS_READ_SMALL_TIME_SERIES.data
+                        };
+                    });
                     await ws.batch.hisDelete(
                         HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
                             batchSize: 12
@@ -140,7 +213,12 @@ describe("client", () => {
 
             describe("delete data greater than batchSize", () => {
                 it("should make more than 1 request", async () => {
-                    ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                    ws.batch.hisRead = sinon.stub().callsFake(() => {
+                        return {
+                            success: HIS_READ_SMALL_TIME_SERIES.data,
+                            errors: []
+                        };
+                    });
                     await ws.batch.hisDelete(
                         HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
                             batchSize: 2
@@ -163,7 +241,12 @@ describe("client", () => {
         describe("option batchSizeEntity", () => {
             describe("entities to delete data smaller than batchSizeEntity", () => {
                 it("should make 1 request", async () => {
-                    ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                    ws.batch.hisRead = sinon.stub().callsFake(() => {
+                        return {
+                            success: HIS_READ_SMALL_TIME_SERIES.data,
+                            errors: []
+                        };
+                    });
                     await ws.batch.hisDelete(
                         HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
                             batchSizeEntity: 10
@@ -178,32 +261,36 @@ describe("client", () => {
 
             describe("entities to delete data greater than batchSizeEntity", () => {
                 it("should make more than 1 request", async () => {
-                    ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                    ws.batch.hisRead = sinon.stub().callsFake(() => {
+                        return {
+                            success: HIS_READ_SMALL_TIME_SERIES.data,
+                            errors: []
+                        };
+                    });
                     await ws.batch.hisDelete(
                         HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
-                            batchSizeEntity: 2
+                            batchSizeEntity: 1
                         }
                     );
-                    expect(ws.hisDelete.calledThrice).to.be.true;
-                    expect(ws.hisDelete.args[0]).to.eql([
-                        HIS_READ_SMALL_TIME_SERIES.ids.slice(0, 2),
-                        "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
-                    ]);
-                    expect(ws.hisDelete.args[1]).to.eql([
-                        HIS_READ_SMALL_TIME_SERIES.ids.slice(2, 4),
-                        "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
-                    ]);
-                    expect(ws.hisDelete.args[2]).to.eql([
-                        HIS_READ_SMALL_TIME_SERIES.ids.slice(4),
-                        "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
-                    ]);
+                    expect(ws.hisDelete.callCount).to.be.equal(5);
+                    for (let i = 0; i < ws.hisDelete.callCount; i++) {
+                        expect(ws.hisDelete.args[i]).to.eql([
+                            [HIS_READ_SMALL_TIME_SERIES.ids[i]],
+                            "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
+                        ], `Index ${i}`);
+                    }
                 });
             });
         });
 
         describe("option returnResult", () => {
             beforeEach(() => {
-                ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                ws.batch.hisRead = sinon.stub().callsFake(() => {
+                    return {
+                        success: HIS_READ_SMALL_TIME_SERIES.data,
+                        errors: []
+                    };
+                });
                 ws.hisDelete = sinon.stub().callsFake(() => ["test123"]);
             });
 
@@ -238,13 +325,27 @@ describe("client", () => {
 
         describe("error handling", () => {
             beforeEach(() => {
-                ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
+                ws.batch.hisRead = sinon.stub().callsFake((...args) => {
+                    return {
+                        success: [],
+                        errors: [{
+                            error: "hisRead error",
+                            args: ["hisRead", ...args]
+                        }]
+                    };
+                });
                 ws.hisDelete = sinon.stub().callsFake(() => {
-                    throw new Error("Test error");
+                    throw new Error("hisDelete error");
                 });
             });
 
-            it("should return encountered errors and arguments used", async () => {
+            it("should return encountered errors and arguments used for hisDelete", async () => {
+                ws.batch.hisRead = sinon.stub().callsFake(() => {
+                    return {
+                        success: HIS_READ_SMALL_TIME_SERIES.data,
+                        errors: []
+                    };
+                });
                 const result = await ws.batch.hisDelete(
                     HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END);
                 expect(result.errors.length).to.equal(1);
@@ -254,8 +355,61 @@ describe("client", () => {
                         HIS_READ_SMALL_TIME_SERIES.ids,
                         "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
                     ],
-                    error: "Test error"
+                    error: "hisDelete error"
                 }]);
+            });
+
+            it("should return encountered errors and arguments used for hisRead", async () => {
+                const result = await ws.batch.hisDelete(
+                    HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END);
+                expect(result.success.length).to.equal(0);
+                expect(result.errors.length).to.equal(1);
+                const { args, error } = result.errors[0];
+                expect(error).to.equal("hisRead error");
+                expect(args[0]).to.equal("hisRead");
+                expect(args[1]).to.eql(HIS_READ_SMALL_TIME_SERIES.ids);
+                expect(args[2].valueOf()).to.equal(TIME_START.valueOf());
+                expect(args[3].valueOf()).to.equal(TIME_END.valueOf());
+            });
+
+            it("should return encountered errors and arguments from both hisRead and hisDelete", async () => {
+                ws.batch.hisRead = sinon.stub().callsFake(() => {
+                    return {
+                        success: [
+                            HIS_READ_SMALL_TIME_SERIES.data[0],
+                            [],
+                            ...HIS_READ_SMALL_TIME_SERIES.data.slice(2)
+                        ],
+                        errors: [
+                            {
+                                error: "hisRead error",
+                                args: [
+                                    "hisRead",
+                                    [HIS_READ_SMALL_TIME_SERIES.ids[1]],
+                                    TIME_START,
+                                    TIME_END
+                                ]
+                            }
+                        ]
+                    };
+                });
+                const { success, errors } = await ws.batch.hisDelete(
+                    HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END);
+                expect(success.length).to.equal(0);
+                expect(errors.length).to.equal(2);
+                const { args: hisReadArgs, error: hisReadError } = errors[0];
+                expect(hisReadError).to.equal("hisRead error");
+                expect(hisReadArgs[0]).to.equal("hisRead");
+                expect(hisReadArgs[1]).to.eql([HIS_READ_SMALL_TIME_SERIES.ids[1]]);
+                expect(hisReadArgs[2].valueOf()).to.equal(TIME_START.valueOf());
+                expect(hisReadArgs[3].valueOf()).to.equal(TIME_END.valueOf());
+                const { args: hisDeleteArgs, error: hisDeleteError} = errors[1];
+                expect(hisDeleteArgs).to.eql([
+                    "hisDelete",
+                    [HIS_READ_SMALL_TIME_SERIES.ids[0], ...HIS_READ_SMALL_TIME_SERIES.ids.slice(2)],
+                    "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
+                ]);
+                expect(hisDeleteError).to.equal("hisDelete error");
             });
         });
     });